requirements.txt

Flask==3.0.0
Flask-Session==0.5.0
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
flask-limiter
flask-wtf
pytz
redis
deepl
ics
ics
ics
ics
libretranslatepy
python-dateutil
apscheduler
python-dotenv
requests
pillow
requests
flask-babel
Flask-Babel>=3.1.0
